+++ changelog.md
@@ -6,6 +6,7 @@
     - TriOS now has a max of 10 API requests active at a time, rather than sending one per mod all at once.
     - Switched to a new HTTP client (wrapper around the native one), so let me know if network calls regress.
     - New setting to control this added to Settings page.
+  - "Cannot write to vmparams file Miko_R3.txt" error when not using JRE 23.
 
 # 0.2.7
 - Added